Update patch set 5
Patch Set 5:

Oh but if there are errors in the code it doesent show them. This problem was on vector because I made a mistake in the json file but test didn't pick it up so that's why jslint and jsonlint were added to fix problem since npm isent running the jslint and jsonlint test.

Patch-set: 5