Update patch set 5
Restored

jshint for users that have not been added to the check +2 test. npm doesent run for users that haven't been added to the list. That's why we should keep the jshint test until we can do npm without infecting or causing damange to the server it is on for all users.

Patch-set: 5
Status: new