Fix  mediawiki-core-bundle-rubocop branch filter
The branch filter for mediawiki-core-bundle-rubocop would never match
the 'master' branch.  The reason is ^(?!xxxxx)$ would end up trying to
match 'master' with '^$' which is obviously False.  Hence the job was
never triggered.

We had the same issue previously with a fundraising job (591297b11a).

Move the start/end of strings markers (^ and $) inside the regex group.
Add a test that covers the job filtering.

Bug: T105178
Change-Id: I9d6946e27333760b7359cfe569fa113723e06b1e