Move some extensions jobs from check-only to check
The mw-checks pipeline will vote +2, which does not make sense when
there are tests that should be run such as jshint.  Those repositories
ended up being in both check and check-voter pipelines which causes
duplicate jobs to be run.

Change some extensions to use mw-checks-test template instead, wich has
the check jobs in the 'check' pipeline instead of 'check-voter'.

Changes:
* ActivityMonitor
* Narayam
* PoolCounter
* SemanticACL
* SemanticCompoundQueries
* SemanticDummyEditor
* SemanticExtraSpecialProperties
* SemanticFormsInputs
* SemanticGenealogy
* SemanticGlossary
* SemanticImageAnnotator
* SemanticInternalObjects
* SemanticPageMaker
* SemanticPageSeries
* SemanticSignup
* SemanticTasks
* SemanticTitle
* SemanticUpdateOnPurge
* WikiLexicalData
* XAnalytics

Change-Id: I8c36c26bca1e59882d40aaa3dcc9888fac1bcfbd