tests: refactor test_repos_have_required_jobs
The test would iterated over:

 pipelines,
 repos regex
 projects,
 via getProjectDef():
   pipelines
   projects
 ..

That makes the test to use up to 50 seconds on my machine.

Refactor the whole test to only load the pipelines and projects job tree
once with the introduciont of getProjectsDefs(). It yields a map of
projects -> pipelines -> jobs (using names, not objects). That then
makes it way fast to run assertion against that single preloaded tree.

Skip asserting check-only and check-voter are in the requirements
pipelines since we detects the requirement by actually looking whether
the project has them.
Slightly refactor the check/test requirements, moving the assertIn
there.

Change-Id: I4a83736fbc43c5b46659ff654add6e38a40ae23a