Update patch set 4
Patch Set 4: Code-Review-1

Since that repo has npm, the jshint job is better replaced by having users to run 'npm test' locally before sending patchsets.

Per Timo in another similar change: the jshint version we run is quite old and could cause a bunch of side effects.  Additionally the use of the template adds jshint jobs in test and gate-and-submit pipelines which are redundant with the npm job.

Patch-set: 4