Update patch set 4
Patch Set 4:

Hi maybe could we open npm to everyone since doing npm locally may be harder for some users that may be in experienced with it or just want to be quick if it isent a majour patch. Or set up another instance that a user can quickly upload the patch to the server would run it and if it detects an error the user can fix it and try again if there are no errors this time a user can download the patch and upload to gerrit. This would not cause security problems since it would be a different instance and nothing to do with your infrastructure.

Patch-set: 4