Update patch set 5
Patch Set 5:

Hi this test also should be done because the updated test I submitted for review removes the old mwembedsupport test. And so this should be here so that the test doesent fail. Or should I do that in the same patch.

Patch-set: 5
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0