Workaround Zuul normalizing approval fields
There is an oddity in Zuul BaseFilter which matches approvals against a
normalized version of the approval description.  It is all lower case.

That causes a Code-Review +2 to no more be recognized because of case
mismatch.

Change gate-and-submit Code-Review+2 approval to use lower case to
workaround Zuul filter.
Add a test covering the event matching.

Bug: T106436
Change-Id: Ia62e5bb91b24f7a33c78ac33b6a633f829b97621