Update patch set 2
Patch Set 2:

(1 comment)

This patch is non sense :-)  Please stop copy pasting random bits and hoping it might do something useful!

There are a few ways to fix the issue of release branches not supporting npm/composer:

A) come back to a job per project. Ie  mwext-SyntaxHighligth_GeSHi-npm  . This way we can set a branch filter to it.  I dont think we want to do that because that consumes a ton of disk space on the slaves and is a pity to maintain.

B) Backport the package.json to the release branches.  That would help back porters but then some checks will fail because code needs to be changed (e.g.: jshint).

C) force merge. Which in this specific case of package.json lacking, sounds acceptable to me.

D) have the npm job to succeed when package.json isn't present, and thus ignore repo that lacks npm configuration.    Maybe we can make it succeed only on REL branches but still failing when it happens to be triggered on master


That is probably worth filling a bug for discussion.

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0