Update patch set 3
Patch Set 3:

Jan as a similar need for the Wikidata repos, that is why I added him as a reviewer.

As I remember, the deployed branch does not have test shipped to reduce the amount of code exposed on the fundraising cluster.  So I think the tests aren't run :-(

For dev branches, you can go with composer that will simply your workflow.  To test the deploy branch, we have to find a creative way.  Maybe the tests should be shipped in the branch but removed on deployment time (rm -fR ./tests ??).

Regardless, we will need two different jobs or a way to switch between composer and vendor repo based on the branch name ($ZUUL_BRANCH).

Patch-set: 3
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0