Update patch set 3
Patch Set 3:

The call to composer was added to the wmff configuration in our branch of civicrm-buildkit in I6b7e01719d90bf3f6450bf8938a6a1a29921aa1c

buildkit is already using composer to install its own dependencies - this patch makes it also use composer to pull in the dependencies for our modules.  Without it we will be unable to test any functionality depending on new or updated packages.

Patch-set: 3