Update patch set 10
Patch Set 10: Code-Review-1

For repositories like this that should not receive a commit unless someone resurrects them, I'd prefer one job that does nothing but always fail.

We should not incur the cost of maintaining the CI for something that should not receive any patches. When someone finds a good reason to resurrect a repository, configuring the CI is just a small part of the needed work.

Patch-set: 10
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1