Update patch set 8
Patch Set 8:

From https://gerrit-review.googlesource.com/Documentation/project-configuration.html :

----------------------------------------------------
Read Only:

The project is read only and all modifying operations on it are disabled. E.g. this means that pushing to this project fails for all users even if they have push permissions assigned on it.

Setting a project to this state is an easy way to temporary close a project, as you can keep all write access rights in place and they will become active again as soon as the project state is set back to Active.

This state also makes sense if a project was moved to another location. In this case all new development should happen in the new project and you want to prevent that somebody accidentally works on the old project, while keeping the old project around for old references.
----------------------------------------------------

Patch-set: 8