Update patch set 7
Patch Set 7: Code-Review-1

(2 comments)

The jobs will not be triggered by zuul with the current configuration. Furthermore I'm not sure, if it's better to introduce a generic job or a job for each extension (you currently doing both). It would be nice to have a task in phabricator to discuss this with the releng/integration developers, first :)

Patch-set: 7
Label: Code-Review=-1