Add job templates to publish Parsoid docs to doc.wikimedia.org
We publish after every merged commit to the master branch in the
source repo.  There is a job to publish from the deploy repo, as well,
in case we ever want to have documentation which reflects the
currently deployed Parsoid.  It is not currently triggered.

This patch also adds `npm run test` to the parsoidsvc-*-npm* jobs,
in order to prevent documentation errors from creeping into our
repo.

Change-Id: I1550106aae658b173783b2921cc01a9a0c611a21