Update patch set 2
Patch Set 2:

If the extension Isent deleted or Isent planed to do it like in the next few days or month you can archive it using the new archive test. Please see paranoid on a example on how to use the test. Archive is meant to fail every time someone uploads to it.

Patch-set: 2