Update patch set 2
Patch Set 2: Code-Review+1

So from what I understand a LocalSettings.php can be placed in the browser tests folder and is appended to the default one. This makes sense to me. Happy to be an early adopter if someone else says this is sane :)

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1