Update patch set 14
Patch Set 14:

The skins and extensions are very similar.  We clone the repositories and run a core testsuite.

Maybe we can factor out the code so it can be used for both extensions and skins?  The macro that clones the dependencies would need to be adjusted so the repos are passed with the parent namespace i.e. we would pass:

 extensions/MobileFrontend

Or:

 skins/Vector

And from there we can rename the 'extensions' jobs to 'integration' or something like that and have them used for both extensions and skins.

For extensions, we run the MediaWiki core PHPUnit test suite named 'extensions'.  The 'skins' suite exists but should include 'structure' and 'suites/LessTestSuite.php'.  I proposed  https://gerrit.wikimedia.org/r/233694 against mediawiki/core to do so.

Patch-set: 14