Fix "curl: Failed writing body"
This happens because 'head' closes the input page from cURL at a point
that cURL does not expect, and cURL will proceed to write more content.

Proxy the stream through tac (which exhausts the full output before
forwarding anything to 'head') to avoid this error.

This will make it slightly slower, but it will make the Jenkins
job output cleaner and less confusing to users.

Change-Id: Iabc4249aa951b71381572dd26241692020f32d19