Update patch set 2
Patch Set 2:

I removed the line disabling jslint because that wasn't what the bug nor your commit message was about.

Removing jslint jobs where it is alread run by npm has the downside of it not running in the check pipeline for users not in the whitelist. So if we want to do that or only run it for check we should do it consistently for all repos.

Patch-set: 2