Update patch set 9
Patch Set 9:

Hi that would be impossible since the js files are from another js project it is from kaultra player meaning that we carnt edit files from another project since it would make it harder to update or edit what we can do is do hacks in it or change code with a well documented change so that a user dosent accidently revert that change.

But we could ignore the js file.

Patch-set: 9