nodepool: placeholder setup script
When preparing the instance, Nodepool require a setup script to be run
on the instance.  We lack sudo rules, so just add a place holder.

The script will be deployed with Trebuchet on the Nodepool server. See
puppet change https://gerrit.wikimedia.org/r/236769

Nodepool invoke the script with the hostname as first argument.

Bug: T111377
Change-Id: I2f053ba5ff8481a4add35ef86cf42a8af2e495e8