Remove quiet option on Android tests
When all tests perform well, Gradle's -q option is great to eliminate
unnecessary noise and shave a second or two off build times.
Unfortunately, tests fail and we need to know why quickly. This patch
removes the quiet option from builds, restoring the default verbosity
level.

For example, prior to this patch, the complete output for a ProGuard
failure was:

  FAILURE: Build failed with an exception.

  * What went wrong:
  Execution failed for task ':app:proguardAlphaReleaseAndroidTest'.
  > java.io.IOException: Please correct the above warnings first.

  * Try:
  Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output.

After this patch, the failure will show that it is due to ProGuard
warnings and what those warnings are.

Change-Id: Ie0805774064b47bbd516094dee84a1075a93691d