Update patch set 1
Patch Set 1: Code-Review-2

Did it in a rush, it has a bunch of unrelated impact on files in the root directory.

That utility is used to generate: https://www.mediawiki.org/wiki/User:Legoktm/ci

Not sure 'dashboard' is the best name since it is very generic, but in the context of integration/config that is probably fine.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2