Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

So all repos having the ruby2.0lint would be considered to be dependent in gate-and-submit.  So a patch to mediawiki/selenium would have to wait for a mediawiki/core change ahead in Zuul queue.


Potentially we could create a mediawiki-core-ruby2.0lint to avoid the problem, this way the mediawiki/selenium would be in its own queue and merge faster.

But we should not bother :-}

Patch-set: 2