[OpenStackManager] Update Jenkins tests
Adds extension-unittests test. It is added to the experiment template for now.

Removes check-only: all test in there and gate-and-submit: all test in
there too.

Test extension-unittests has all those tests in it plus maybe a few more,

Add composer-test

Add phplint until extension-unittests is ready to be in template:

Change-Id: I65079a8890fae99a425bea976a3ad34a91ba5416