Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

Yeah set -u is an excellent choice :-}

Since that overrides Jenkins default set, you will want to add set -e.

The DOC_* variables are not injected if they are empty string.  But ZUUL_UUID is a default parameter to the job, and default to an empty string. So we need to keep the [ -z $ZUUL_UID ] test.

Patch-set: 2
Label: Code-Review=-1