Update patch set 1
Patch Set 1: Code-Review-1

Doh I looked at phpcs as a non serious / pet project and never came back to it.  And now we are about to make it voting?   Great.

An issue though is that it is going to fail on past release branch. And I think we don't need it on wmf branches.

So you want a branch filter:

 - name: mediawiki-core-phpcs:
   branch: (?!^wmf/.*|^REL1_23|^REL1_24|^REL1_25|^fundraising/REL)


Note the various branch filters are rather messy, but that one should do.

Patch-set: 1
Label: Code-Review=-1