Update patch set 2
Patch Set 2:

> And I think we don't need it on wmf branches.

We *do* need it on wmf/ branches given that it also checks for things like closing PHP tags, BOMs, and whitespace before the opening <?php (which has caused an outage in the past).

However, I don't want to backport a bunch of stuff to make the current wmf branch pass, so we should exclude it for now, and in a week make it vote on wmf/ branches.

Patch-set: 2
Label: Verified=0