Test mediawiki gate queue only has mw projects
Detect non MediaWiki projects that share jobs with MediaWiki related
projects. That is to prevent the projects from being merged in the
'mediawiki' queue of gate-and-submit.

A project get merged in the 'mediawiki' change queue either because it
shares a job with one of the mediawiki repositories or when it shares a
job with a project that got unintendly merged in. In the later case,
they are not triggering an error.

To run the test:

tox -e py27 --
tests/test_zuul_scheduler.py:TestZuulScheduler.test_only_mediawiki_projects_in_mediawiki_gate
--logging-level=ERROR

Bug: T107529
Change-Id: Icd3903d1a275677fdfc42c341db2e7dbe06d275e