Update patch set 6
Patch Set 6:

I guess the source repositories should have a change that pass npm/composer.  Then we can merge the CI configuration and +2 all the pending patches.  They should then merge in.

In this case, some of those repositories/changes are not passing :-(  So I should not have deployed the CI change cause the source repositories are now blocked.

Patch-set: 6