Update patch set 2
Patch Set 2:

The composer update stuff is currently a macro in integraiont/config.git/jjb/..., we could make it a target in the makefile and then use the makefile from a jjb macro as an intermediate step. To move to this.

One important question: When using parallelization with make, does it cleanly separate the output of each command that runs in parallel, so that one can easily read the output of one command as a continuous set of lines?

Patch-set: 2