Experiment YAML aliases and template inheritance
The Jenkins jobs have been made generic to be triggered by any
repository. A side effect is Zuul dependent pipeline (gate-and-submit)
handles merging almost all of the projects in the same queue
'mediawiki'.

As a result, a change unrelated to MediaWiki such as for
integration/config ends up being queued behind mediawiki/core changes
that takes a while to run.

This experiment propose to split some jobs again, but instead of
prefixing them with the repository name, prefix them with the queue they
are intended for.

JJB job-templates support default values for parameters when they are
defined inside the job template. That let you invoke the job-template
without defining the variable.

YAML has support for merging mappings and then override given values.
Abuse that feature to craft a back compatibility job 'phpunit' with an
empty queue.  That let us avoid copy pasting job-templates that solely
vary on the name ('phpunit' vs '{queue}phpunit').

Change-Id: I17c650acce623c3a1356dfe57ab591335c2cd169