Update patch set 2
Patch Set 2: Code-Review-1

This repo is included in the Wikidata build, thus it has nothing to do with vendor.git. At least until we move from having the Wikidata build and vendor.git to one having only one. The idea of this patch is to fix the phpunit jobs on this repo, as it is currently broken.

Anyway there are a few things that need fixing here before its ready: These jobs may not be in check. The zend variant should only be run during gate and during the zend queue. Then the extension-unittests template needs to be removed, but the non phpunit jobs from it added to this repo.

Patch-set: 2
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1