Remove mediawiki-core-release job
The job was supposed to craft a tarball of MediaWiki core and run tests
for us whenever a tag is released.  We decided end of March 2015 to use
a manual process instead (T94583).

Remove reference to mediawiki-core-release and the related test.

That removes a coupling with Gerrit replication (/srv/ssd/gerrit) which
is not even available on labs slaves.

Bug: T86661
Bug: T94583
Change-Id: I06a4454f6eb32280cd21d65f9386e2365ee2db05