tests: restore failures report at end of run
Whenever a failure occurs, we have to dig for it in somewhere in the
output.  The issue is caused by nose-timer which causes the failure
message to be outputed asap instead of being delayed at the end of the
test run.

Remove nose-timer dependencies and --with-timer.  That will get the
failures to be buffered and written at the end of the run.

Bug: T107683
Change-Id: I5144dd2036d19270d947ffea31932eebce449438