Instruct bundler to install under vendor/bundle
bundler defaults to install to GEM_HOME which on Debian is a system
directory.  It thus attempt to sudo to copy the gems which is a no/no on
our Nodepool instances (jenkins user has no sudo rights).

Instead instruct bundler to install gems to vendor/bundle which is the
default behavior when using --deployment.

Update job rake-jessie

Bug: T114860
Change-Id: I32ac084cbd2100594275991cca8add815f282f18