Update extension-unittests tests
Add tests mwext-testextension-zend-composer and mwext-testextension-hhvm-composer to experiment:

This will help us decide which extensions need to use these tests it is
already added to extension-unittests-generic so adding it to
extension-unittests.

Probaly need to do something like

mwext-{name}-testextension-zend-composer

and

mwext-{name}-testextension-zend-composer

But not sure how to do that.

Change-Id: I4a821f089effca8562d1e134851ccc19ccd378d7