Use pipeline name as context for Zuul diff
Reviewing diff generated by integration-zuul-layoutdiff is quite
annoying since we have no idea which pipelines are being changed. GNU
diff has the ability to display the context based on a regular
expression, albeit it is limited to 40 characters.

Hijack the layout output:

-INFO:zuul.IndependentPipelineManager:Configured Pipeline Manager test
+Pipeline: test

Use GNU diff option --show-function-line='^Pipeline: ' to teach it that
is our context.

Change-Id: Iade99387875b78477f22166098ce97f0c7c56278