Remove redundant Android tests
The Android Wikipedia app is tested in two phases: assembly and JUnit
JVM tests, and Android instrumentation tests. The latter requires that
the Android emulator be in a settled state which is a long running task
performed in parallel with the first phase. This patch makes use of a
new Gradle target in the second phase, connectedAlphaReleaseAndroidTest.
Previously, JVM JUnit tests were executed redundantly by the all
encompassing testAllAlphaRelease target.

Change-Id: I85fc0721ae97a7390aaf532b69415b0f25ef1fa3