Allow skins to also be tested like extensions can
This test is for skins

It allows for skins to be tested just like extensions can.

Adds metrolook and vector as dependencies. They are disabled by default
since they arn't enabled by default in layout.xml. We will enable this
after this patch is merged to test weather this works.

Bug: T68926
Change-Id: I1be4454b1aa7935ec2e9ac3703eee3907e5101e8