Remove non-voting jobs from gate-and-submit
There is no need to have non-voting in gate-and-submit since they do not
participate in the final score and are never going to prevent a merge.
They are already run in test for user feedback.

Remove:

mediawiki/extensions/CirrusSearch  mwext-CirrusSearch-whitespaces
mediawiki/extensions/VisualEditor  commit-message-validator
operations/debs/wikistats          debian-glue-non-voting

Add a test to enforce it.

Change-Id: Idcbe01988e54581b3c70a3159547fc51734968e9