Update patch set 3
Patch Set 3:

Legoktm: Yes I will remove phplint but jshint even though it is a different version is for non-whitelisted users or the user will only be able to run the jsonlint test.

phplint is run through composer-test but not for non-whitelisted users. so if there is an error in the code a user may not notice it until a whitelisted user rechecks it.

Patch-set: 3