Update patch set 2
Patch Set 2:

> I disagree that VisualEditor browser tests were fine before we deleted the jobs.

No, but they were fine (some working, some not, in the slow process of being fixed despite being foisted upon the team) before you broke them.

> I have asked on several places (qa mailing list, phabricator, gerrit) before deleting the jobs and got no reply, so I have assumed nobody cared.

None of these are appropriate venues for synchronous or semi-synchronous communication. IRC exists. E-mail exists. Silence is not consent.

> Reverting this change would show that all Internet Explorer and Safari jobs are broken because of T114362.

OK, so… why is that task not, for instance, a high-priority tell-all-engineers task? Why is it that the first I hear that the CI infrastructure is totally broken for the most important targets a comment in this commit? We already trivially test everything in Chrome/Firefox as part of normal development; the main value of browser tests are IE, Safari and particularly iOS Safari…

> As a first step in recreating the jobs, I will recreate one Firefox job and see what needs to be done to make it green.

Thank you.

Patch-set: 2
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0