Update patch set 8
Patch Set 8: Code-Review-1

We changed the operations-puppet-doc templates between rebased of this change. It now fetch code under $WORKSPACE/src unlike the vagrant job.

The best would probably to reuse the job template 'operations-puppet-doc' and covert it to to a generic one that can be used by MWV.  Drop the macro in the process (just inline it in the job template).

Patch-set: 8
Label: Code-Review=-1