Run Ruby jobs using Rake for mediawiki/core
Replace mediawiki-core-bundle-rubocop by rake-jessie. And thus delete
the job from Jenkins.

Use Zuul skip-if to prevent the rake-jessie job from running on old and
fundraising branches.  Note skip-if branch filtering is a blacklist
where as branch: filter is a white list, so the regex logic is reversed.

Adjust integration test.

Bug: T114860
Change-Id: I41c4755d05ef07f93b7acf319bf0ecb452f3722c