Remove Android dependency workaround
The SDK Manager Plugin previously did not properly download all
dependencies[0]. The latest version does, so the workaround of using the
Jenkins Android Emulator Plugin to download missing dependencies is no
longer needed. This should give us a nice performance boost for the
apps-android-wikipedia-publish job and reduce the load on jobs truly
dependent on the AndroidEmulator node.

[0] https://github.com/JakeWharton/sdk-manager-plugin/issues/57

Change-Id: Idade13b6505eeb3f7154881e3e970db0f783e9ff
Depends-On: Ia0f5f8dc64a676bdead0bd79b36b8071d45cb516