Update patch set 2
Patch Set 2:

Niedzielski I gave this patch a quick look but haven't carefully looked at it.  

The job is triggered on timer but its scm still refer to Zuul parameters. So the git url is wrong:

 stderr: fatal: '$ZUUL_URL/$ZUUL_PROJECT' does not appear to be a git repository

Maybe we can instead use the scm polling system which is built-in the Jenkins git plugin. An example is mw-tools-scap-tox-doc-publish which has:

    triggers:
      - pollscm: '@hourly'
    scm:
      - git:
          url: 'https://phabricator.wikimedia.org/diffusion/MSCA/scap'
          branches:
            - 'origin/master'
          wipe-workspace: true

Copy paste and replace the URL with the Gerrit one and it should work.

There is a poll log available:

https://integration.wikimedia.org/ci/job/mw-tools-scap-tox-doc-publish/scmPollLog/

And it should only build when a new commit happened :-}

Patch-set: 2