Mark mwext-.*-jslint as deprecated
The jobs run both jsonlint and jshint. We definitely want jsonlint to
run on all repos but since jshint does not pass we can not make them
voting.

We should get rid of those jobs entirely in favor of using the npm entry
point.  First step is to have Zuul report back in Gerrit the job is
deprecated, a strategy I used previously for yamllint in 5e78f7537.

Change-Id: I7b58f62a7f7ab427bf55782c50672d8cad5f292f