[BlueSpiceExtensions] just run php/json linters
BlueSpiceExtensions is a collection of various extensions. If we wanted
to test it, we would want a dedicated job that loads all of them and
then run the various test framework we have.

Hence, the jobs running on it are currently useless. Thus disabling
tests and just run voting phplint/jsonlint jobs.

Change-Id: I835397945d2ce27f857993181714c909911bb40b