Add CirrusSearch and Elastica as dependencies for Wikibase tests
since we are implementing hook handlers for Cirrus and
working on better integration with Cirrus / ElasticSearch.

afaik, Elastica provides enough abstraction that allows
us to have tests for this integration w/o actually having
ElasticSearch installed. Integration tests that do need
ElasticSearch can also be run on the Wikimedia CI.

Also, GeoData was missing as a dependency for the
Wikidata build, so added that as well.

Change-Id: Ief8b7770fa79b39730442d54cf06a33c80182ebc