Enable submodules for operations/mediawiki-config phpunit tests
This is required for I6743721f. The mediawiki/event-schemas submodule
it being added to mediawiki-config to remove duplication of the schemas
across multiple repositories. To pass the tests this submodule
needs to be initialized.

As far as i can tell from digging through other config and the unit tests
run on this repository, the correct method is to copy the scm: configuration
from the parent job (use-remote-zuul-shallow-clone) but remove the part
that disables submodules.

Change-Id: I27baf00f091567090bc766d47d1674ff291dcf6e