Enable submodules for operations/mediawiki-config phpunit tests
This is required for I6743721f. The mediawiki/event-schemas submodule
it being added to mediawiki-config to remove duplication of the schemas
across multiple repositories. To pass the tests this submodule
needs to be initialized.

This has been tested against https://gerrit.wikimedia.org/r/#/c/255135/
and sets up the submodules appropriately.

Change-Id: I27baf00f091567090bc766d47d1674ff291dcf6e