Enable submodules for operations/mediawiki-config phpunit tests
This is required for I6743721f. The mediawiki/event-schemas submodule
it being added to mediawiki-config to remove duplication of the schemas
across multiple repositories. To pass the tests this submodule
needs to be initialized.

Override the scm for operations-mw-config-phpunit to have the Git
plugin:
- wipe the workspace
- use a shallow clone (speed up)
- explicity define we want submodules to be processed

This has been tested against https://gerrit.wikimedia.org/r/#/c/255135/
and sets up the submodules appropriately.

Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: I27baf00f091567090bc766d47d1674ff291dcf6e