nodepool: run apt-get upgrade while snapshoting
The snapshots are created out of a base image that might have outdated
package by the time we create snapshots. The base image is only updated
from time to time.

An example is packages in puppet are usually marked with ensure =>
present and they end up being frozen to the version we had when build
the reference image.

Make setup_node.sh to run apt-get upgrade to ensure snapshots are up to
date.

Bug: T120961
Change-Id: I59fb3af690f445cc3c07f369796ea6f2fef33eb9