If one batch fails for WPT, fail the whole job.
Before we didn't look at individual returns for each batch job we
run in Jenkins for WebPageTest, so if one failed we stopped and
didn't run the rest. We now have moved the queueing to the
internal WebPageTest queue (for speedup and scaling) and make the
code handle errors and now we also catch errors in the Jenkins and
sums the return codes and if one batch run fails, we get one error
in Jenkins (and can act on it in the future).

Bug: T120365
Change-Id: I3c519ef8d28e31122f917516440c493536b627a4