Update patch set 4
Patch Set 4:

Well we can hold on this until Flow developers stabilize the tests on their local machine?   There is little point in having a job spamming errors constantly, that is just going to distract you :-}

I proposed to restore them merely because we use to have an IE job.  Win 8.0 / IE 10 might need to be reconsidered as well.

Patch-set: 4