Assert gate jobs are properly configured
In the JJB templates for mediawiki-extensions-* jobs, we define a list
of extensions participating and the jobs must be triggered by all
projects.

There is nothing preventing to add the trigger to Zuul without updating
the Jenkins jobs.  Nor is there anything preventing to add a dependency
in the job which is not triggering it.

Introduce a test to assert JJB <-> Zuul are in sync.

CirrusSearch has been triggering the gate though it is not even defined
as a dependency in the Jenkins jobs. That has been covered by 297dff88.

List repositories in Zuul having extension-gate and assert the set is
equal to the list of dependencies in mediawiki-extensions-* jobs.

Change-Id: I9dc6e4ada3fd0686428a0b6cc42db07d4eeac60c