Update patch set 1
Patch Set 1:

NSFileRepo uses a specific job "mwext-NSFileRepo-testextension-zend" which does inject LockDown as a dependency.   The CI loader loads the repo triggering the patch first, then its dependencies.

Looks like only NSFileRepoHooks::onUserCan() rely on lockdown and gracefully skip the userCan whenever the function is not found.

Maybe you can do the check via the 'setupAfterCache' hook. At that point MediaWiki should have loaded all extensions. So a NSFileRepoHooks::setupAfterCache() can raise an exception or just log a warning that the permissions are not going to be handled properly because lockdown is missing.

---

As for this change, the dependencies in zuul/ext_dependencies.py are only applied on the generic jobs triggered by various repos (such as 'mwext-testextension-zend'.    In this case the repo triggers a more specific job "mwext-NSFileRepo-testextension-zend"  which has the dependency hardcoded in the job definition.

Patch-set: 1