Update patch set 1
Patch Set 1:

Glad you like it. Sorry for the annoyance, but in the end that means people can load both extensions in whatever order and it will just work ™.  Should make the installation easier.

Another thing, is when the repositories are migrated to extension registration ( https://www.mediawiki.org/wiki/Manual:Extension_registration ) I don't think the load order might not be guaranteed, but you can mention dependencies and check whether another ext is loaded.    But that migration should be probably be for some other change :D

Patch-set: 1