Update patch set 2
Patch Set 2:

I gave it a try on gallium and triggered https://gerrit.wikimedia.org/r/#/c/259167/21

The job run https://integration.wikimedia.org/ci/job/mwext-testextension-zend/18821/ is with Wikidata being injected in the Jenkins job.

There is a few test errors https://integration.wikimedia.org/ci/job/mwext-testextension-zend/18821/testReport/


A big trouble is that adding the dependencies also get the tests for all of them to be run as well.  Maybe we can change the Jenkins job to only run the test of the triggering extension.

Patch-set: 2