[Graph] add VisualEditor as a dependency
Graph conditionally register additional resources whenever VisualEditor
is present. The CI job was missing VE causing Graph to not register the
resources and thus the structure suite would not verify them.

Add VisualEditor as a dependency to Graph so it can eventually register
additional modules.

Notes: depends on load order.

Bug: T122872
Change-Id: I2f2c8f208499a3bd8435c74c052e110a7a595ae2