Dummy job to let wikimedia/fundraising/crm gate
On wikimedia/fundraising/crm , the gate-and-submit might end up being
empty which causes Zuul to idle and not merge the change.

The jslint/phplint might not trigger due to file filtering and
wikimedia-fundraising-civicrm is not run on deployment branch,
we still need a job to run for Zuul to gate-and-submit.

Add noop job to test and gate-and-submit pipelines.

Bug: T120881
Change-Id: I0c3a4139fd1590e9dec7acd1a85247ecfb2a3825