Update patch set 3
Patch Set 3:

Maybe we could create a separate test that is a non voting job that does what this patch is trying to do. It would allow the us to see which repo needs tests added but not cause the test to fail.

Patch-set: 3
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0