Lets install MySQL before installing extension and extensions dependencies
Because some extensions requires sql updates,
And so as MySQL is installed after extensions are installed then after that the sql updates happened,
MySQL fails because the sql changes are needed.

So lets install MySQL before extensions then install the extensions then run the database update.

The non generic tests already does this so that's why it works. Lets update these tests to do it too.


How it was set prevous was that it would install the mediawiki extensions before installing MySQL then running MySQL update. The problem this caused was that some mediawiki extensions require that they run MySQL first so by installing MySQL before extensions this stop the web updater throwing errors and allows us to run MySQL update. This is already done in non generic test. Just need to add it to generic tests which this patch does.

Change-Id: I5f202d1f3adcd7d7f2641719c2e636830c3425fd