Lets install MySQL before installing extensions dependencies
Because some extensions requires sql updates,
And so as MySQL is installed after extensions are installed then after that the sql updates happened,
MySQL fails because the sql changes are needed.

So lets install MySQL before extensions then install the extensions then run the database update.

The non generic tests already does this so that's why it works. Lets update these tests to do it too.

Change-Id: I5f202d1f3adcd7d7f2641719c2e636830c3425fd