Move SemanticFormsInputs back to mediawiki-extensions.yaml file
It is currently failing in the ext_dependencies.py file because the
dependacy extension is loaded after the actual extension requiring it so
it fails because the main extensions loads first looking for the dependacy
extension but carnt find it.

This is temporarily until extension.json is created for this repo.

Change-Id: Ib1e1f3d2340dbe97721864005a90dbf8ca240a02